### 2024-05-01 #177 @whit5224
Dolphot-LC Software Submission
Submitting Author: Ramona White (@whit5224)
All current maintainers: (@whit5224, @patkel)
Package Name:  Dolphot-LC
One-Line Description of Package: Dolphot-LC is an automated Hubble Space Telescope (HST) data pipeline based on the popular Dolphot analysis package; this package allows for the creation of lightcurves and difference images from HST data.
Repository Link:  https://github.com/patkel/dolphot_lc
Version submitted:   1
Version accepted: TBD

---

- [ X ] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package after should it be accepted.
- [ X ] I have read and will commit to package maintenance after the review as per the [pyOpenSci Policies Guidelines][Commitment].

Dolphot-LC is an automated Hubble Space Telescope (HST) data pipeline based on the popular Dolphot analysis package. This package allows for the creation of lightcurves and difference images from HST data. Dolphot-LC allows the user to run their own fits images through the pipeline and generate results and requires a coadded template image and science images that are already aligned to template. To make package use easier, Dolphot-LC has been documented extensively in a Read the Docs: https://dolphot-lc.readthedocs.io/en/latest/index.html and Jupyter Notebook: https://nbviewer.org/gist/whit5224/287af111f44bf83a23eaaf19a5121c75.

	- [ X ] Data extraction
	- [ X ] Data processing/munging

- [ X ] Astropy:[My package adheres to Astropy community standards](https://www.pyopensci.org/software-peer-review/partners/astropy.html)

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience are scientist and astrophysicists interested in analyzing HST and JWST data with the goal of creating light curves and difference images.

Dolphot-LC implements and expands upon the DOLPHOT stellar photometry package. Our software makes it possible apply DOLPHOT to difference images in order to measure changes in brightness due to astrophysical transients. Dolphot-LC improves quality and usability by allowing the creating of light curves and difference images.

- [ X ] does not violate the Terms of Service of any service it interacts with.
- [ X ] uses an [OSI approved license][OsiApprovedLicense].
- [ X ] contains a README with instructions for installing the development version.
- [ X ] includes documentation with examples for all functions.
- [ X ] contains a tutorial with examples of its essential functions and uses.
- [ X ] has a test suite.
- [ X ] has continuous integration setup, such as GitHub Actions CircleCI, and/or others.

- [ X ] Yes I am OK with reviewers submitting requested changes as issues to my repo. Reviewers will then link to the issues in their submitted review.

- [ X ] I have read the [author guide](https://www.pyopensci.org/software-peer-review/how-to/author-guide.html).
- [ X ] I expect to maintain this package for at least 2 years and can help find a replacement for the maintainer (team) if needed.

- [ X ] [Last but not least please fill out our pre-review survey](https://forms.gle/F9mou7S3jhe8DMJ16). This helps us track submission and improve our peer review process. We will also ask our reviewers
