### 2024-05-04 #177 @Batalex
Hey @whit5224, thank you for submitting Dolphot-LC! I am Alex, pyOpenSci current Editor-in-Chief. I went through the docs and the code, and I think that Dolphot-LC is in scope for us.
Over the next few days, I'll go over the E-i-C checks so that we can address a few things before moving on to the actual review.
