### 2024-05-19 #177 @Batalex
Hi there! Thank you for submitting your package for pyOpenSci review. Below are the basic checks that your package needs to pass to begin our review. If some of these are missing, we will ask you to work on them before the review process begins.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

---
---

Hey there,
Sorry it took me so long to get back to you.

I think that `dolphot_lc` needs some more work before we proceed with the review. I made the call to not check some of the items, here is why.

> The package imports properly into a standard Python environment `import package`.

The package installation does not install its dependencies. This is problematic because the user has to go to the repository to find the `requirements.txt` file. I encourage you to take a look at our [packaging guide](https://www.pyopensci.org/python-package-guide/) to properly set up that. On a side note, I am curious as to how you built the wheel hosted on [pypi](https://pypi.org/project/dolphot-lc/#files) since it's not clear from the CI and the source code.

> Documentation

I'd like to see more content to get a better grip of what `dolphot_lc` does (arguably, I'm not familiar with the field, so this might be a ME problem). Namely, think of your documentation as something that should fill the two following axes:
- expertise level: tutorial, how-to guide, technical reference
- content structure: **what** are we going to do, **how** are we going to do it using your package, **why** is it done that way?

> Readme

Think of the readme as the front page for your package. Like a resume, you only have a few moments of attention span from a reader before they lose interest and look for something else.
Here is an example you can take inspiration from

<details><summary>Expand</summary>

````md

# dolphot_lc

A Python package for <stuff>.

A brief description of what this project does and who it's for.

Install `dolphot_lc` with pip

pip install dolphot_lc
```

## Getting started

The 30 seconds tutorial.

```py
import antigravity
```

[Documentation](https://linktodocumentation)

## Contributing

Contributions are always welcome!

See `contributing.md` for ways to get started.

Please adhere to this project's `code of conduct`.

````

Using our package guide, the readme will also be present on the [pypi front page](https://pypi.org/project/dolphot-lc/#description)

> Contributing

The file is present, but the content is lacking. For instance, you set the expectations that users can contribute by fixing bugs or proposing new features, but there is no explanation as to how to do that (how to build the package, how to test it, etc.)

> Tests

Same as before, I encourage you to take a look at our guide. Using a proven test framework will increase the quality of your tests, as well as aligning `dolphot_lc` on the community-established practices.
