### 2024-07-05 #177 @cmarmo
Hi @whit5224 , I'm Chiara and I'm following up your submission as editor in chief.
Do you mind giving an update about the initial checks on your code?
No pressure, just let us know how do you plan to proceed so we can fairly share the packages among editors and reviewers.
Thank you for submitting your package to pyOpenSci!
