### 2025-02-21 #177 @whit5224
Hi there, thank you for checking in! I am sorry it has been so long, this
has flown under our radar. We still want to proceed with this review, and
aim to have the changes suggested by Alex done within the month.

On Fri, Feb 21, 2025 at 2:38 PM Leah Wasser ***@***.***>
wrote:

> hi, there @whit5224 <https://github.com/whit5224>, I'm just checking in
> on this review, given it's been open for some time, and it's been 9 months
> since we've heard from you! I understand that life gets busy, and that is
> OK!
>
> We just need to understand if you have the bandwidth to proceed with this
> review or not. We have two options!
>
>    1. If you have a timeline within the next month or so, we can move
>    forward once you have made the changes requested by Alex
>    2. If you feel like you don't have the bandwidth, we can close this
>    issue (just for now) and reopen it in the future or create a new one when
>    you have the bandwidth!
>
> Please let us know how you'd like to proceed and if we can support you in
> any way with moving this forward.
>
> I hope that you are doing well and look forward to your response.
>
