### 2025-05-27 #177 @dhomeier
> This should have astropy label. FYI. Thanks!

@pllim apparently I had removed the label right after first adding it last year already, and just after you had entered the package into the "migrating" packages list in astropy/astropy.github.com#571, but I can't remember why I would have done that...
