### 2025-10-07 #177 @lwasser
Hi friends. This review has been quiet since May! I'm just checking in here. Are we waiting for an Astropy editor to have bandwidth to move it forward?
